fibb and simple fix

+++ test/Program.cs
@@ -1,12 +1,12 @@
 ﻿using FuncsWithNumbers;
 
 namespace test
-{
+{   
     internal class Program
     {
         static void Main(string[] args)
         {
-            List<int> num = NumbersTasks.GenerateNumbers(NumbersTasks.IsFibonacci(), 100);
+            List<long> num = NumbersTasks.GenerateNumbers(NumbersTasks.IsFibonacci(), 100);
             foreach (var item in num)
             {
                 Console.WriteLine(item);
